+++ bauh/gems/appimage/__init__.py
@@ -5,7 +5,7 @@ from bauh.api.constants import CONFIG_PATH
 from bauh.commons import resource
 
 ROOT_DIR = os.path.dirname(os.path.abspath(__file__))
-LOCAL_PATH = '{}/.local/share/bauh/appimage'.format(Path.home())
+LOCAL_PATH = '{}/.local/share/bauh/appimage'.format(str(Path.home()))
 INSTALLATION_PATH = LOCAL_PATH + '/installed/'
 SUGGESTIONS_FILE = 'https://raw.githubusercontent.com/vinifmor/bauh-files/master/appimage/suggestions.txt'
 CONFIG_FILE = '{}/appimage.yml'.format(CONFIG_PATH)